svn path=/nixpkgs/branches/stdenv-updates/; revision=19786

+++ pkgs/build-support/fetchgit/builder.sh
@@ -7,7 +7,9 @@ if test -n "$rev"; then
   cd $out
   git checkout $rev
 fi
-find $out -name .git\* | xargs rm -rf
+
+if test -z "$leaveDotGit"; then
+    find $out -name .git\* | xargs rm -rf
+fi
 
 stopNest
-
